redirect root route to '/dashboard' if user is signed in
add route guard to redirect root route to '/dashboard' if user is signed in
