+++ apps/client/src/app/app.routes.ts
@@ -5,15 +5,20 @@ import { LandingPageComponent } from './landing-page/landing-page.component';
 import { GoogleCallbackComponent } from './auth/callbacks/google-callback.component';
 import { GithubCallbackComponent } from './auth/callbacks/github-callback.component';
 import { SetupAccountComponent } from './setup-account/setup-account.component';
-import { SignedInGuard } from './routeGuards/signed-in.guard';
+import { AuthGuard } from './routeGuards/auth.guard';
 import { DashboardComponent } from './dashboard/dashboard/dashboard.component';
+import { NoAuthGuard } from './routeGuards/no-auth.guard';
 
 export const routes: Routes = [
-    { path: '', component: LandingPageComponent },
+    {
+        path: '',
+        component: LandingPageComponent,
+        canActivate: [NoAuthGuard],
+        data: { redirectTo: '/dashboard' },
+    },
     {
         path: 'auth',
         component: AuthComponent,
-        // canActivate: [AuthGuard],
         children: [
             { path: 'verify-email', component: VerifyEmailComponent },
             { path: 'callback/google', component: GoogleCallbackComponent },
@@ -23,11 +28,11 @@ export const routes: Routes = [
     {
         path: 'setup-budget',
         component: SetupAccountComponent,
-        canActivate: [SignedInGuard],
+        canActivate: [AuthGuard],
     },
     {
         path: 'dashboard',
         component: DashboardComponent,
-        canActivate: [SignedInGuard],
+        canActivate: [AuthGuard],
     },
 ];
