+++ apps/client/src/app/routeGuards/no-auth.guard.ts
@@ -2,13 +2,13 @@ import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
 import Session from 'supertokens-web-js/recipe/session';
 
-export const SignedInGuard: CanActivateFn = async (route, state) => {
+export const NoAuthGuard: CanActivateFn = async (route, state) => {
     const router = inject(Router);
 
-    if (await Session.doesSessionExist()) {
+    if (!(await Session.doesSessionExist())) {
         return true;
     }
 
-    router.navigate(['/auth']);
-    return false;
+    const redirectTo = route.data['redirectTo'] || '/auth';
+    return router.parseUrl(redirectTo);
 };
